update(search): determine caps using search types instead of cats (#976)
The main change is using the search caps for `EPISODE`, `SEASON` and
`MOVIE` MediaTypes in `indexerDoesSupportMediaType`. A movie tracker has
a single tv category of miniseries which causes `cross-seed` to search
all episodes and season packs. Using the `tvsearch` caps would address
this issue instead of using cats. This seems to work well with all
indexers.

I added a migration for audio, music, and book search caps for
completeness. They are used in addition to categories for these media
types. I also added a new column for name so we can log the names. This
is updated on each torznab results call to avoid using the Prowlarr and
Jackett APIs.

When we are updating indexer caps (startup and job), we will log the
indexer mediatypes support in verbose to make the behavior more clear.