ehancement(qbit): use includeTrackers includeFiles for info api (#985)
`includeTrackers` is a new param added in qbit 5.1.0 that allows us to
get the trackers for all torrents with the single `/torrents/info`
endpoint. Previously, we would need to send an individual api request
for each torrent, but wasn't a big deal since it only does that for
first time users and the `cleanup` job.

`includeFiles` does not exist, but if it were to it would be handled the
same.