enhancement(snatch): don't retry frequently failed recent snatches (#986)
This adds a limit to the retries for snatches in the past day. Snatches
will still be tried up to 5 times with a delay in between, but if all
those fail, if this snatch fails again in the next day we won't retry.
We will still always attempt it once.