fix(SelectField): actually updates with initial value
For an unknown-to-me reason, the initial value was not updating the
`select-field` `Select` inputs. Found a slightly hacky way to get it to
work by adding a `key` prop. Many people have the same issue with shadcn
on GitHub and Stackoverflow, this solution was the one that seemed to
work best