fix(linkDir): use stat.dev for choosing linkDir only if all are unique (#991)
If not all `linkDirs` have a unique `stat.dev`, we will choose one by
using the fallback method of trying to link on all of them. Using
`stat.dev` is simply a less intensive check if it's available.

Also logging `stat.dev` for each `linkDir` and `dataDir` on startup to
make debugging easier.