enhancement(inject): add injected candidates to searchee pool (#1007)
This in support of this
[FAQ](https://www.cross-seed.org/docs/basics/faq-troubleshooting#my-partial-matches-from-related-searches-are-missing-the-same-data-how-can-i-only-download-it-once)
for rss/announce/inject. Since these process once candidate at a time,
we can add the injected candidate for the searchee pool for the next
candidate.

For search/webhook, this does not apply. Since these process one
searchee at a time and all valid candidates are injected, there is no
value to be derived unless the newly injected torrents recheck+complete
in a couple seconds.