perf(inject): up to 16x speed up for job by reducing api calls (#1028)
For `cross-seed inject`, we check every searchee against the .torrent
file and run the action stage for all of them. This is because one
searchee might have a file that another doesn't that the candidate
needs. e.g:
* Candidate: mkv, srt, nfo
* SearcheeA: mkv, srt
* SearcheeB: mkv, nfo

The issue is that a single candidate can easily have 10+ valid searchees
that the user already has. The slowdown happens when we are in the
action stage for each injection, we call `client.inject()` which first
checks if the torrent exists in client to return `ALREADY_EXISTS`. This
API call can take a bit if the client has a large amount of torrents.

Instead, we already know if the candidate already exists in client as we
need to use that save path if it does, rather than any from `linkDirs`.
By removing this extra exists check, I was able to reduce the time from
`232s` to `14s` for injecting `10` .torrent files. This is because the
first searchee injects the candidate then all others link the new files
and recheck if necessary.