update(gh/README) update and simplify README.md file on github (#1033)
This PR updates the README to have more flowing language, removes the
reference to the "New breaking changes in v6" for a gentle mention of
"if you're migrating", and removes autotorrent entirely - as we now
support all of the clients AT2 supported, and it's of no benefit.

I also added a Repobeats graph at the bottom to gauge everyone's
thoughts on the new look. You can see it at the link below.


https://github.com/cross-seed/cross-seed/tree/update-gh-readme

<hr>
Update our "description" to "Fully-automatic cross-seeding"

- [x] documentation
- [x] unraid template
- [ ] README.md (this)