fix(perf): reduce memory usage during reverse lookup (#1036)
#1026 seems to have increase the occurrence of #1015. This usually
happens during `announce` and to users with a large amount of torrents
or data entries. #1026 mitigated this by only allowing a single
candidate from `rss` to be processed at a time but `announce` was left
unchecked. We will now globally only allow a single reverse lookup by
moving the `AsyncSemaphore` into `checkNewCandidateMatch`. This only
covers the reverse lookup portion as the decide stage is unlikely to be
the cause.