enhancement(torrent_cache): save entries with mediaType and tracker (#1042)
This means when a user runs `cross-seed restore` after losing their
torrent data, the injected torrents will be linked to `linkDir/Tracker`
instead of `linkDir/UnknownTracker` for all.

The format is now the same as the one used for `outputDir` exception
with `.cached`, `[mediaType][tracker]name[infoHash].cached.torrent`. The
`.cached` is keep when copying to `outputDir` so they can still be
distinguished from regular outputDir` entries.

This change is only future facing, it will not modify existing entries
in the cache unless it's seen again. Instead, when a user runs
`cross-seed restore`, it uses the updated format for the new entries and
applies `[tracker]` to them when copying based on the tracker hosts in
the .torrent files. So if the users has at least one of the new format
for each tracker, we should be able to rebuild the entire filename
metadata for all torrents.

Before we were only saving the infoHash in the cache, which means all
entries were the same size. But now since we are saving the name, we can
run into max path length issues, so now for both `outputDir` and
`torrent_cache`, it will truncate the .torrent title if the name is too
long (`255 bytes` for total path). It's also now easier to find torrents
in the cache since it has the name and tracker.

We will now also prune the `decision` table for invalid entries with the
cleanup job. First we remove all entries with a `null` infoHash as we no
longer save those. We also check each infoHash to see if it's still
present in the torrent cache. This should ensure the number of unique
infoHashes in the table is the same as the number of .torrent files in
the cache.

Cleaning the decision table is rarely needed so it's done roughly once
per month using `Math.random()`. If the `cleanup` job is triggered from
`/api/job`, it will always be done.