enhancement(db/indexer): store tracker hosts for restore (#1045)
We will now store the tracker hosts from the latest snatch in the
indexer table so we can map arbitrary .torrent files with the right
indexer name. This allows us to avoid using `UnknownTracker` for
`cross-seed restore` and `cross-seed inject`.

With this change, the implementation of #1042 has also been updated. We
can prune the decision table with every run of the cleanup job since we
are no longer need any I/O operations aside from database lookups.